+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.3.0
 * Add package#is_empty method
 * Add package#exists method
+* Add package#is_built method
 
 ### 1.2.0
 * Limit memory consumption when downloading large package file